weather-app/requirements/prod.txt
celery[redis]==5.3.6
celery-redbeat==2.2.0
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flower==1.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
